Moved Glavenus configs
