+++ BNA_KC_Vehicles/Light/Glavenus/CfgPatches.hpp
@@ -1,0 +1,25 @@
+class CfgPatches
+{
+    class BNA_KC_Vehicles_Glavenus
+    {
+        addonRootClass= "BNA_KC_Vehicles";
+        author = "SweMonkey and DartRuffian";
+        requiredVersion = 0.1;
+        requiredAddons[] =
+        {
+            "BNA_KC_Vehicles",
+                // Core Config
+            "A3_Soft_F_MRAP_02",
+                // Karetel
+            "BNA_KC_Vehicles_Weapons",
+                // Co-Ax
+            "BNA_KC_Gear_Infantry"
+                // Unit
+        };
+        units[] =
+        {
+
+        };
+        weapons[] = {};
+    };
+};
